requirements.txt
beautifulsoup4==4.14.3
brotli==1.2.0
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
fonttools==4.61.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pillow==12.1.0
pycparser==2.23
pycups==2.0.4
pydyf==0.12.1
pyphen==0.17.2
requests==2.32.5
soupsieve==2.8.1
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.6.2
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
